Refactor repository and model structures for improved clarity and consistency
- Updated return types in Azure and Postgres repositories to use `crate::db::models::BlockHistogramData`.
- Refactored LocalRepository to utilize `ScheduleId` instead of `i64` for schedule identifiers, enhancing type safety.
- Adjusted various methods in LocalRepository, AnalyticsRepository, and VisualizationRepository to accommodate the new `ScheduleId` type.
- Simplified constraints and scheduling block models by removing unnecessary conversions to quantity types.
- Enhanced the clarity of the `Schedule` and `SchedulingBlock` models by providing clear documentation for internal use.
- Updated service functions to align with the new model structures, ensuring consistent data handling across the application.
- Modified tests to reflect changes in model structures and ensure correctness in the new implementations.

+++ rust_backend/src/services/visibility.rs
@@ -125,7 +125,7 @@ pub fn compute_visibility_histogram_rust(
             }
         }
 
-        bin.visible_count = visible_blocks.len() as u32;
+        bin.visible_count = visible_blocks.len() as i64;
     }
 
     Ok(bins)
@@ -181,7 +181,7 @@ mod tests {
     #[test]
     fn test_compute_histogram_single_block() {
         use crate::api::Period;
-        use crate::siderust::astro::ModifiedJulianDate;
+        use crate::models::ModifiedJulianDate;
 
         let block = BlockHistogramData {
             scheduling_block_id: 1,
@@ -206,7 +206,7 @@ mod tests {
     #[test]
     fn test_compute_histogram_priority_filter() {
         use crate::api::Period;
-        use crate::siderust::astro::ModifiedJulianDate;
+        use crate::models::ModifiedJulianDate;
 
         let blocks = vec![
             BlockHistogramData {
@@ -240,7 +240,7 @@ mod tests {
     #[test]
     fn test_compute_histogram_overlapping_periods() {
         use crate::api::Period;
-        use crate::siderust::astro::ModifiedJulianDate;
+        use crate::models::ModifiedJulianDate;
 
         // Same block with multiple overlapping periods in same bin
         let block = BlockHistogramData {
